Former-commit-id: ac38ea98743bc2a01cdcc14498cf9e26cf2e2c2e

+++ cmd/platonecli/packet/tx_calls.go
@@ -6,7 +6,7 @@ import (
 
 // InnerCallCommon
 // extract the common part of all the inner calls
-func InnerCallCommon(funcName string, funcParams []string, txType int) *InnerCall {
+func InnerCallCommon(funcName string, funcParams []string, txType uint64) *InnerCall {
 	// parse the function parameters
 	funcName, funcParams = utl.FuncParse(funcName, funcParams)
